Mark function.parameters as required
In https://github.com/openai/openai-node/issues/181#issuecomment-1611712836, a user claimed that a function without `parameters` specified will not be called, but the same function with `parameters: { 'type': 'object', properties: {} }` is called. This leads me to believe that the user should always pass `parameters`, even if they must be an empty object.

A better approach to marking this required might be to make this the default on the server side, so that users do not need to pass this object for functions which accept no parameters.